ADD: A new `<Status>` component to display the status of a runner formatted
